backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-generativeai
pandas
pandas_ta
pydantic
python-multipart
redis>=5.0.0
async-lru
beautifulsoup4
gunicorn
msgpack
fyers-apiv3
websockets
httpx